[SCR-673] Update perform feature so that features are specified in locales
Previously, we allowed a feature only if it was for the USA.
Now, we check the locale file for that country for the feature.feature_name.
Unless the feature is set and equal to true, the feature cannot be run.
Default truevalues can be set in the en.yml locale.